# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  dto.CommentRequest:
    properties:
      content:
        type: string
      parent_id:
        type: integer
    required:
    - content
    type: object
  dto.CommentResponse:
    properties:
      content:
        type: string
      created_at:
        type: string
      id:
        type: integer
      parent_id:
        type: integer
      user:
        $ref: '#/definitions/dto.ReviewUserInfo'
    type: object
...
  dto.CreateReviewRequest:
    properties:
      content:
        type: string
      game_id:
        type: integer
      img:
        type: string
      is_spoiler:
        type: boolean
      recommend:
        enum:
        - recommend
        - mixed
        - not_recommend
        type: string
      title:
        type: string
    required:
    - content
    - game_id
    - recommend
    type: object
...
  dto.ErrorResponse:
    properties:
      code:
        type: string
      error:
        type: string
      field:
        type: string
    type: object
...
  dto.PaginatedResponse-array_dto_ReviewTrendingResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/dto.ReviewTrendingResponse'
        type: array
      pagination:
        $ref: '#/definitions/dto.PaginationDTO'
      status:
        type: string
    type: object
...
  dto.PaginationDTO:
    properties:
      current_page:
        type: integer
      limit:
        type: integer
      total_pages:
        type: integer
      total_records:
        type: integer
    type: object
...
  dto.ReviewGameInfo:
    properties:
      id:
        type: integer
      release_date:
        type: string
      thumbnail:
        type: string
      title:
        type: string
    type: object
...
  dto.ReviewTrendingResponse:
    properties:
      comment_count:
        type: integer
      content:
        type: string
      created_at:
        type: string
      game:
        $ref: '#/definitions/dto.ReviewGameInfo'
      is_spoiler:
        type: boolean
      like_count:
        type: integer
      rating:
        type: number
      review_id:
        type: integer
      user:
        $ref: '#/definitions/dto.ReviewUserInfo'
    type: object
  dto.ReviewUserInfo:
    properties:
      avatar:
        type: string
      id:
        type: integer
      username:
        type: string
    type: object
...
  dto.SuccessResponse-array_dto_CommentResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/dto.CommentResponse'
        type: array
    type: object
...
  dto.SuccessResponse-dto_CommentResponse:
    properties:
      data:
        $ref: '#/definitions/dto.CommentResponse'
    type: object
...
  dto.SuccessResponse-dto_ReviewTrendingResponse:
    properties:
      data:
        $ref: '#/definitions/dto.ReviewTrendingResponse'
    type: object
...
  dto.SuccessResponse-string:
    properties:
      data:
        type: string
    type: object
...
  dto.UpdateReviewRequest:
    properties:
      content:
        type: string
      img:
        type: string
      is_spoiler:
        type: boolean
      recommend:
        enum:
        - recommend
        - mixed
        - not_recommend
        type: string
      title:
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: Hệ thống API cho Gaming Box.
  title: Gaming Box API
  version: "1.0"
paths:
...
  /reviews:
    post:
      consumes:
      - application/json
      description: Tạo một review cho game (Cần đăng nhập)
      parameters:
      - description: Thông tin review
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.CreateReviewRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.SuccessResponse-dto_ReviewTrendingResponse'
      security:
      - BearerAuth: []
      summary: Tạo Review mới
      tags:
      - Reviews
  /reviews/{id}:
    delete:
      description: Xóa một review (Chỉ chủ sở hữu)
      parameters:
      - description: ID của Review
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.SuccessResponse-string'
      security:
      - BearerAuth: []
      summary: Xóa Review
      tags:
      - Reviews
    put:
      consumes:
      - application/json
      description: Chỉnh sửa review đã viết (Chỉ chủ sở hữu mới có quyền)
      parameters:
      - description: ID của Review
        in: path
        name: id
        required: true
        type: integer
      - description: Thông tin cập nhật
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateReviewRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.SuccessResponse-dto_ReviewTrendingResponse'
      security:
      - BearerAuth: []
      summary: Cập nhật Review
      tags:
      - Reviews
  /reviews/{id}/comments:
    get:
      description: Lấy tất cả bình luận của một review
      parameters:
      - description: ID của Review
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.SuccessResponse-array_dto_CommentResponse'
      summary: Lấy danh sách bình luận
      tags:
      - Reviews
    post:
      consumes:
      - application/json
      description: Bình luận vào một review (Cần đăng nhập)
      parameters:
      - description: ID của Review
        in: path
        name: id
        required: true
        type: integer
      - description: Nội dung bình luận
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.CommentRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.SuccessResponse-dto_CommentResponse'
      security:
      - BearerAuth: []
      summary: Thêm bình luận
      tags:
      - Reviews
...
  /reviews/trending:
    get:
      description: Lấy danh sách review phổ biến có phân trang
      parameters:
      - description: Trang hiện tại (Mặc định 1)
        in: query
        name: page
        type: integer
      - description: Số lượng mỗi trang (Mặc định 10)
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.PaginatedResponse-array_dto_ReviewTrendingResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Lấy danh sách Review thịnh hành
      tags:
      - Reviews
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"